use usize

+++ core/src/dynmod.rs
@@ -36,7 +36,7 @@ fn score(env: &Env, str: String, query: String) -> Result<Option<Vector>> {
     }
     let _inner_vec: Vec<i32> = _vec.unwrap();
     let mut vec = env.make_vector(_inner_vec.len(), ())?;
-    let index: i32 = 0;
+    let index: usize = 0;
     _inner_vec.into_iter().for_each(|data| {
         vec.set(index, data);
         index += 1;
